Snap/ restore with infinite loop...
