+++ src/wasm/CMakeLists.txt
@@ -9,7 +9,6 @@ set(LIB_FILES
         FunctionLoader.cpp
         LocalFunctionLoader.cpp
         FileserverFunctionLoader.cpp
-        WasmModuleRegistry.cpp
         S3FunctionLoader.cpp
         IRModuleRegistry.cpp
         instance.cpp
